wip on header hash in hdk
